requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic>=2.0
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
aiofiles
jinja2
python-dateutil 